chore: add use cases to usecases layer
