+++ app/src/main/kotlin/swc/microservice/truck/usecases/CreateTruck.kt
@@ -1,0 +1,7 @@
+package swc.microservice.truck.usecases
+
+import swc.microservice.truck.entities.Truck
+
+class CreateTruck(private val truck: Truck) : TruckUseCase<String> {
+    override fun execute(manager: TruckManager): String = manager.createTruck(truck)
+}
