+++ app/src/main/kotlin/swc/microservice/truck/usecases/DeleteTruck.kt
@@ -1,0 +1,7 @@
+package swc.microservice.truck.usecases
+
+class DeleteTruck(private val truckId: String) : TruckUseCase<Unit> {
+    override fun execute(manager: TruckManager) {
+        manager.deleteTruck(truckId)
+    }
+}
