+++ app/src/main/kotlin/swc/microservice/truck/usecases/GetTruck.kt
@@ -1,0 +1,7 @@
+package swc.microservice.truck.usecases
+
+import swc.microservice.truck.entities.Truck
+
+class GetTruck(private val truckId: String) : TruckUseCase<Truck> {
+    override fun execute(manager: TruckManager): Truck = manager.getTruck(truckId)
+}
