+++ app/src/main/kotlin/swc/microservice/truck/usecases/TruckCount.kt
@@ -1,0 +1,6 @@
+package swc.microservice.truck.usecases
+
+class TruckCount : TruckUseCase<Int> {
+
+    override fun execute(manager: TruckManager): Int = manager.getTruckCount()
+}
