+++ app/src/main/kotlin/swc/microservice/truck/usecases/TruckNextId.kt
@@ -1,0 +1,5 @@
+package swc.microservice.truck.usecases
+
+class TruckNextId : TruckUseCase<String> {
+    override fun execute(manager: TruckManager): String = "Truck${manager.getTruckCount()}"
+}
