+++ app/src/main/kotlin/swc/microservice/truck/usecases/TruckUseCase.kt
@@ -1,0 +1,5 @@
+package swc.microservice.truck.usecases
+
+interface TruckUseCase<T> {
+    fun execute(manager: TruckManager): T
+}
