--- app/src/main/kotlin/swc/microservice/truck/usecases/queries/CreateTruck.kt
@@ -1,8 +1,0 @@
-package swc.microservice.truck.usecases.queries
-
-import swc.microservice.truck.entities.Truck
-import swc.microservice.truck.usecases.TruckManager
-
-class CreateTruck(private val truck: Truck) : TruckQuery<String> {
-    override fun execute(manager: TruckManager): String = manager.createTruck(truck)
-}
