--- app/src/main/kotlin/swc/microservice/truck/usecases/queries/TruckCount.kt
@@ -1,8 +1,0 @@
-package swc.microservice.truck.usecases.queries
-
-import swc.microservice.truck.usecases.TruckManager
-
-class TruckCount : TruckQuery<Int> {
-
-    override fun execute(manager: TruckManager): Int = manager.getTruckCount()
-}
