--- app/src/main/kotlin/swc/microservice/truck/usecases/queries/TruckNextId.kt
@@ -1,7 +1,0 @@
-package swc.microservice.truck.usecases.queries
-
-import swc.microservice.truck.usecases.TruckManager
-
-class TruckNextId : TruckQuery<String> {
-    override fun execute(manager: TruckManager): String = "Truck${manager.getTruckCount()}"
-}
