--- app/src/main/kotlin/swc/microservice/truck/usecases/queries/TruckQuery.kt
@@ -1,7 +1,0 @@
-package swc.microservice.truck.usecases.queries
-
-import swc.microservice.truck.usecases.TruckManager
-
-interface TruckQuery<T> {
-    fun execute(manager: TruckManager): T
-}
