+++ app/src/test/kotlin/swc/microservice/truck/TruckUseCasesTest.kt
@@ -1,0 +1,4 @@
+package swc.microservice.truck
+
+class TruckUseCasesTest {
+}
